requirements.txt
# run this command to install the requirements: 
# pip install -r requirements.txt

flet==0.84.0
Pillow==12.1.1  # PIL: OCR / previews + Convert Image (any-to-any format conversion)
pytesseract==0.3.13
keyboard==0.13.5
pyperclip==1.11.0
deep-translator==1.11.4
python-dotenv==1.0.1
pyttsx3==2.99
pywin32==311
lmdb==1.8.1
Flask==3.1.3
qrcode==8.0
PyMuPDF==1.25.3
watchdog==6.0.0
scrapy==2.13.3
scrapy-playwright==0.0.44
playwright==1.54.0

sockets==1.0.0
# Upload tab → Mobile: Bluetooth (Windows WinRT)
winrt-runtime==3.2.1
winrt-Windows.Devices.Bluetooth==3.2.1
winrt-Windows.Devices.Bluetooth.Rfcomm==3.2.1
winrt-Windows.Devices.Enumeration==3.2.1
winrt-Windows.Foundation==3.2.1
winrt-Windows.Foundation.Collections==3.2.1
winrt-Windows.Networking==3.2.1
winrt-Windows.Networking.Sockets==3.2.1
winrt-Windows.Storage.Streams==3.2.1
